requirements.txt
requests
pytest
python-telegram-bot
